+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/drive/opmode/DriveVelocityPIDTuner.java
@@ -15,6 +15,7 @@ import com.acmerobotics.roadrunner.profile.MotionProfileGenerator;
 import com.acmerobotics.roadrunner.profile.MotionState;
 import com.acmerobotics.roadrunner.util.NanoClock;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.RobotLog;
@@ -50,6 +51,8 @@ import java.util.List;
  */
 @Config
 @Autonomous(group = "drive")
+@Disabled
+
 public class DriveVelocityPIDTuner extends LinearOpMode {
     public static double DISTANCE = 72; // in
