+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/drive/opmode/FollowerPIDTuner.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.team26396.roadrunner.drive.opmode;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.team26396.roadrunner.drive.SampleMecanumDrive;
@@ -23,6 +24,8 @@ import org.firstinspires.ftc.team26396.roadrunner.trajectorysequence.TrajectoryS
  */
 @Config
 @Autonomous(group = "drive")
+@Disabled
+
 public class FollowerPIDTuner extends LinearOpMode {
     public static double DISTANCE = 48; // in
