+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/drive/opmode/TrackingWheelLateralDistanceTuner.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.team26396.roadrunner.drive.opmode;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.util.Angle;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.RobotLog;
@@ -63,6 +64,8 @@ import org.firstinspires.ftc.team26396.roadrunner.drive.StandardTrackingWheelLoc
  */
 @Config
 @TeleOp(group = "drive")
+@Disabled
+
 public class TrackingWheelLateralDistanceTuner extends LinearOpMode {
     public static int NUM_TURNS = 10;
