+++ TeamCode/src/main/java/org/firstinspires/ftc/team26396/roadrunner/drive/opmode/TurnTest.java
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.team26396.roadrunner.drive.opmode;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.team26396.roadrunner.drive.SampleMecanumDrive;
@@ -11,6 +12,8 @@ import org.firstinspires.ftc.team26396.roadrunner.drive.SampleMecanumDrive;
  */
 @Config
 @Autonomous(group = "drive")
+@Disabled
+
 public class TurnTest extends LinearOpMode {
     public static double ANGLE = 90; // deg
